+++ apps/main-app/ts/features/authentication/activeSessionLogin/screens/cie/ActiveSessionLoginCieConsentDataUsageScreen.tsx
@@ -25,8 +25,10 @@ import {
 } from "../../../common/analytics/cieAnalytics";
 import { AUTH_ERRORS } from "../../../common/components/AuthErrorComponent";
 import { AUTHENTICATION_ROUTES } from "../../../common/navigation/routes";
-import { onLoginUriChanged } from "../../../common/utils/login";
-import { originSchemasWhiteList } from "../../../common/utils/originSchemasWhiteList";
+import {
+  onLoginUriChanged,
+  originSchemasWhiteList
+} from "../../../common/utils";
 import { LoaderComponent } from "../../shared/components/LoaderComponent";
 import { ACS_PATH } from "../../shared/utils";
 import {
