+++ apps/main-app/ts/features/authentication/login/cie/screens/CieConsentDataUsageScreen.tsx
@@ -23,8 +23,10 @@ import {
 } from "../../../common/analytics/cieAnalytics";
 import { AUTHENTICATION_ROUTES } from "../../../common/navigation/routes";
 import { loginFailure, loginSuccess } from "../../../common/store/actions";
-import { onLoginUriChanged } from "../../../common/utils/login";
-import { originSchemasWhiteList } from "../../../common/utils/originSchemasWhiteList";
+import {
+  onLoginUriChanged,
+  originSchemasWhiteList
+} from "../../../common/utils";
 
 export type CieConsentDataUsageScreenNavigationParams = {
   cieConsentUri: string;
